requirements.txt
fastapi==0.116.1
uvicorn==0.15.0
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
alembic==1.16.4
python-dotenv==1.1.1
python-multipart==0.0.20
passlib==1.7.3
bcrypt==3.2.0
python-jose==3.5.0
apscheduler>=3.11.0
email-validator>=2.2.0
itsdangerous>=2.1.2
